move logic generate path into entity Media
