+++ src/services/storage.js
@@ -6,37 +6,16 @@ const shortHash = require('short-hash');
 const s3Config = {
   bucket: 'media-on-demand',
   region: 'us-east-1',
+  accessKeyId: '',
+  secretAccessKey: ''
 };
 const s3 = new S3(s3Config);
-
-function _uniquePath({
-  tenant,
-  url,
-  preset = 'default',
-  width = 'auto',
-  height = 'auto'
-}) {
-  let hash = shortHash(url);
-  let filename = path.basename(url);
-
-  return `media/${tenant}/${hash}/${preset}/${width}/${height}/${filename}`;
-}
-
-function _originalPath({
-  tenant,
-  url,
-}) {
-  let hash = shortHash(url);
-  let filename = path.basename(url);
-
-  return `media/${tenant}/${hash}/${filename}`;
-}
 
 function meta(media, original = false) {
   return new bluebird((resolve, reject) => {
     let key = original ?
-      _originalPath(media) :
-      _uniquePath(media);
+      media.originalPath :
+      media.uniquePath;
 
     s3.headObject({
       Bucket: s3Config.bucket,
@@ -54,8 +33,8 @@ function meta(media, original = false) {
 function get(media, original = false) {
   return new bluebird((resolve, reject) => {
     let key = original ?
-      _originalPath(media) :
-      _uniquePath(media);
+      media.originalPath :
+      media.uniquePath;
 
     media._request = s3.getObject({
       Bucket: s3Config.bucket,
